% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetConstructionEvaluation.R
\docType{data}
\name{ai3ConceptSets}
\alias{ai3ConceptSets}
\title{AI3 Concept Sets}
\format{
a dataframe with 681 rows and 2 variables
\itemize{
\item id : The ID of the concept set
\item conceptId:  The concept ID of a concept included in the concept set
}
}
\usage{
data(ai3ConceptSets)
}
\description{
Concept sets as created by AI3 in the MindsMeetMachines challenge. The dataset includes the set of concept IDs
included in each of the six concept sets according to AI3.
}
\keyword{datasets}
