% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetConstructionEvaluation.R
\docType{package}
\name{ConceptSetConstructionEvaluation-package}
\alias{ConceptSetConstructionEvaluation}
\alias{ConceptSetConstructionEvaluation-package}
\title{ConceptSetConstructionEvaluation: Evaluating Concept Set Contructors}
\description{
Evaluating concept set contructors using various gold standards.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ohdsi.github.io/DatabaseConnector/}
  \item \url{https://github.com/OHDSI/DatabaseConnector}
  \item Report bugs at \url{https://github.com/OHDSI/DatabaseConnector/issues}
}

}
\author{
\strong{Maintainer}: Martijn Schuemie \email{schuemie@ohdsi.org}

Authors:
\itemize{
  \item Martijn Schuemie \email{schuemie@ohdsi.org}
  \item Azza Shoaibi
  \item Patrick Ryan
}

}
\keyword{internal}
