% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EvaluateConceptSets.R
\name{evaluateConceptSets}
\alias{evaluateConceptSets}
\title{Evaluate concept sets}
\usage{
evaluateConceptSets(conceptSets)
}
\arguments{
\item{conceptSets}{The concept sets to evaluate. A data frame with two columns: \code{id}: the ID of the concept set, and
\code{conceptId}: the concept ID of the concept included in the concept set. Can contain 1 or more rows
per concept set.}
}
\value{
The evaluation of the concept sets
}
\description{
Evaluate concept sets
}
\examples{
data(ai3ConceptSets)
results <- evaluateConceptSets(ai3ConceptSets)
results$f1ConservativeWeighted[results$id == "Total"]

}
