% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConceptSetTargets.R
\name{getConceptSetTargets}
\alias{getConceptSetTargets}
\title{Get the concept set targets included in the concept sets}
\usage{
getConceptSetTargets()
}
\value{
A data frame with one row per concept set, and the following columns:
\itemize{
\item \strong{id}: A unique identifier. Use this when feeding the results in \code{evaluateConceptSets()}.
\item \strong{name}: The name of the clinical idea for which to generate the concept set.
\item \strong{definition}: A long free-text definition of the clinical idea, including the intended use of the concept set.
\item \strong{closestConceptId}: The concept ID of the concept that most closely matched the clinical idea.
\item \strong{closestConceptName}: The name of the concept that most closely matched the clinical idea.
}
}
\description{
Get the concept set targets included in the concept sets
}
\examples{
getConceptSetTargets()

}
