% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjudicationEvaluation.R
\name{getConceptsForAdjudication}
\alias{getConceptsForAdjudication}
\title{Get concepts for concept adjudication evaluation}
\usage{
getConceptsForAdjudication()
}
\value{
A data frame with the following columns:
\itemize{
\item \strong{targetName}: The name of the target for the concept set (what the concept set should be about).
\item \strong{targetDefinition}: A detailed description of the concept set target.
\item \strong{conceptId}: The concept ID of the candidate concept.
\item \strong{conceptName}: The name of the candidate concept.
\item \strong{vocabularyId}: The vocabulary ID of the candidate concept.
\item \strong{domainId}: The domain ID of the candidate concept.
\item \strong{conceptClassId}: The concept class ID of the candidate concept.
}
}
\description{
Get a set of concept set targets and candidate concepts for evaluating the process of concept adjudication
(determining) whether a concept belongs to a concept set.

Use the \code{evaluateConceptAdjudication()} function to evaluate your adjudications of these concepts.
}
