% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AdjudicationEvaluation.R
\name{evaluateConceptAdjudication}
\alias{evaluateConceptAdjudication}
\title{Evaluate concept adjudication}
\usage{
evaluateConceptAdjudication(concepts)
}
\arguments{
\item{concepts}{The data frame returned by \code{getConceptsForAdjudication()} with an extra 'adjudication' column, having
value 'YES' if the concept should be included in a concept set for the target, or 'NO' otherwise.}
}
\value{
A data frame with performance statistics.
}
\description{
Evaluate concept adjudication
}
\examples{
concepts <- getConceptsForAdjudication()

# Using random adjudication for this example:
concepts$adjudication <- sample(c("YES", "NO"), nrow(concepts), replace = TRUE)

evaluateConceptAdjudication(concepts)

}
